Completed first step of porting existing code into a new structure.
