--- Linux/Plex-HD.json
@@ -1,133 +1,0 @@
-{
-  "PresetList" : [
-    {
-      "AlignAVStart" : true,
-      "AudioCopyMask" : [
-        "copy:aac",
-        "copy:ac3"
-      ],
-      "AudioEncoderFallback" : "ca_aac",
-      "AudioLanguageList" : [
-
-      ],
-      "AudioList" : [
-        {
-          "AudioBitrate" : 160,
-          "AudioCompressionLevel" : -1,
-          "AudioDitherMethod" : "auto",
-          "AudioEncoder" : "ca_aac",
-          "AudioMixdown" : "stereo",
-          "AudioNormalizeMixLevel" : false,
-          "AudioSamplerate" : "auto",
-          "AudioTrackDRCSlider" : 0,
-          "AudioTrackGainSlider" : 0,
-          "AudioTrackQuality" : -1,
-          "AudioTrackQualityEnable" : false
-        },
-        {
-          "AudioBitrate" : 640,
-          "AudioCompressionLevel" : -1,
-          "AudioDitherMethod" : "auto",
-          "AudioEncoder" : "copy:ac3",
-          "AudioMixdown" : "5point1",
-          "AudioNormalizeMixLevel" : false,
-          "AudioSamplerate" : "auto",
-          "AudioTrackDRCSlider" : 0,
-          "AudioTrackGainSlider" : 0,
-          "AudioTrackQuality" : -1,
-          "AudioTrackQualityEnable" : false
-        }
-      ],
-      "AudioSecondaryEncoderMode" : true,
-      "AudioTrackSelectionBehavior" : "first",
-      "ChapterMarkers" : true,
-      "ChildrenArray" : [
-
-      ],
-      "Default" : false,
-      "FileFormat" : "av_mp4",
-      "Folder" : false,
-      "FolderOpen" : false,
-      "InlineParameterSets" : false,
-      "Mp4HttpOptimize" : true,
-      "Mp4iPodCompatible" : false,
-      "PictureAutoCrop" : true,
-      "PictureBottomCrop" : 12,
-      "PictureChromaSmoothCustom" : "",
-      "PictureChromaSmoothPreset" : "off",
-      "PictureChromaSmoothTune" : "none",
-      "PictureCombDetectCustom" : "",
-      "PictureCombDetectPreset" : "default",
-      "PictureDARWidth" : 0,
-      "PictureDeblockCustom" : "strength=strong:thresh=20:blocksize=8",
-      "PictureDeblockPreset" : "off",
-      "PictureDeblockTune" : "medium",
-      "PictureDeinterlaceCustom" : "",
-      "PictureDeinterlaceFilter" : "decomb",
-      "PictureDeinterlacePreset" : "default",
-      "PictureDenoiseCustom" : "",
-      "PictureDenoiseFilter" : "off",
-      "PictureDenoisePreset" : "light",
-      "PictureDenoiseTune" : "none",
-      "PictureDetelecine" : "off",
-      "PictureDetelecineCustom" : "",
-      "PictureForceHeight" : 0,
-      "PictureForceWidth" : 0,
-      "PictureHeight" : 0,
-      "PictureItuPAR" : false,
-      "PictureKeepRatio" : true,
-      "PictureLeftCrop" : 0,
-      "PictureLooseCrop" : false,
-      "PictureModulus" : 2,
-      "PicturePAR" : "auto",
-      "PicturePARHeight" : 27,
-      "PicturePARWidth" : 32,
-      "PictureRightCrop" : 0,
-      "PictureRotate" : "angle=0:hflip=0",
-      "PictureSharpenCustom" : "",
-      "PictureSharpenFilter" : "off",
-      "PictureSharpenPreset" : "medium",
-      "PictureSharpenTune" : "none",
-      "PictureTopCrop" : 8,
-      "PictureWidth" : 0,
-      "PresetDescription" : "",
-      "PresetName" : "Plex-HD",
-      "SubtitleAddCC" : false,
-      "SubtitleAddForeignAudioSearch" : true,
-      "SubtitleAddForeignAudioSubtitle" : false,
-      "SubtitleBurnBDSub" : true,
-      "SubtitleBurnBehavior" : "foreign",
-      "SubtitleBurnDVDSub" : true,
-      "SubtitleLanguageList" : [
-
-      ],
-      "SubtitleTrackSelectionBehavior" : "none",
-      "Type" : 1,
-      "UsesPictureFilters" : true,
-      "UsesPictureSettings" : 2,
-      "VideoAvgBitrate" : 9000,
-      "VideoColorMatrixCodeOverride" : 0,
-      "VideoEncoder" : "x264",
-      "VideoFramerate" : "30",
-      "VideoFramerateMode" : "pfr",
-      "VideoGrayScale" : false,
-      "VideoLevel" : "4.0",
-      "VideoOptionExtra" : "",
-      "VideoPreset" : "slow",
-      "VideoProfile" : "high",
-      "VideoQSVAsyncDepth" : 4,
-      "VideoQSVDecode" : false,
-      "VideoQualitySlider" : 20,
-      "VideoQualityType" : 2,
-      "VideoScaler" : "swscale",
-      "VideoTune" : "",
-      "VideoTurboTwoPass" : false,
-      "VideoTwoPass" : true,
-      "x264Option" : "",
-      "x264UseAdvancedOptions" : false
-    }
-  ],
-  "VersionMajor" : 42,
-  "VersionMicro" : 0,
-  "VersionMinor" : 0
-}
