+++ connections.py
@@ -9,7 +9,6 @@ Date                : 09/01/2023
 =======================================================================================
 Version History     :
 09/01/2023          : Initial Release
-
 """
 
 ## Standard library imports
